+++ nope/lib/assets/js/content.js
@@ -17,7 +17,7 @@
         .state('app.content.detail', {
           url: '/view/{id:int}',
           views: {
-            'content@app.content': {
+            'content': {
               templateUrl: function($stateParams) {
                 return 'view/' + $stateParams.contentType + '/detail.html'
               },
@@ -55,20 +55,6 @@
       $scope.contentType = $stateParams.contentType;
       $scope.contentsList = [];
       $scope.q = $location.search();
-      $scope.selection = [];
-
-      $scope.select = function(c,i) {
-        if($rootScope.nope.isIframe) {
-          var callerScope = $nopeUtils.getContentModalCallerScope();
-          $scope.selection = callerScope.selectedItem(c);
-          callerScope.$apply();
-        } else {
-          $state.go('app.content.detail', {
-            id:c.id,
-            contentType: $stateParams.contentType
-          });
-        }
-      }
 
       $scope.search = function(q, page) {
         page = page || 1;
@@ -179,13 +165,14 @@
         id: $stateParams.id
       }, function(data) {
         $scope.content = data;
-        $scope.$parent.selectedContent = $scope.content;
+        $scope.$parent.$parent.selectedContent = $scope.content;
       });
 
       $scope.$on('nope.content.updated', function(e, data) {
+        console.log(data.id, $stateParams.id);
         if(data.id === $stateParams.id) {
           $scope.content = data;
-          $scope.$parent.selectedContent = $scope.content;
+          $scope.$parent.$parent.selectedContent = $scope.content;
         }
       });
     }])