+++ nope/lib/views/gallery/form.php
@@ -1,4 +1,4 @@
-<form name="galleryForm" ng-submit="save();">
+<form name="galleryForm" ng-submit="$parent.save(gallery);">
   <div class="panel panel-default">
     <div class="panel-heading content-author" ng-if="gallery.id">
       <div class="btn-group btn-group-xs toolbar pull-right">
@@ -15,31 +15,31 @@
         <div class="col col-md-9">
           <div class="form-group" ng-class="{'has-error':(!galleryForm.slug.$valid && galleryForm.slug.$touched)}">
             <label class="control-label">Slug</label>
-            <input type="text" name="slug" class="form-control" ng-model="gallery.slug" required ng-pattern="<?php echo \Nope\Utils::SLUG_REGEX_PATTERN; ?>" ng-trim="false" />
+            <input type="text" name="slug" class="form-control" ng-model="gallery.slug" required ng-pattern="<?php echo \Nope\Utils::SLUG_REGEX_PATTERN; ?>" ng-trim="false" placeholder="Gallery slug" />
           </div>
         </div>
         <div class="col col-md-3">
           <div class="form-group">
             <label class="control-label">Priority</label>
-            <input type="text" name="slug" class="form-control" ng-model="gallery.priority" ng-pattern="/^[0-9]+$/" />
+            <input type="text" name="slug" class="form-control" ng-model="gallery.priority" ng-pattern="/^[0-9]+$/" placeholder="Priority: higher first" />
           </div>
         </div>
       </div>
       <div class="form-group">
-        <label>Body</label>
-        <textarea name="body" class="form-control" ng-model="gallery.body"></textarea>
+        <label>Description</label>
+        <textarea name="body" class="form-control" ng-model="gallery.body" placeholder="Gallery description" rows="5"></textarea>
       </div>
       <div class="form-group">
         <label>Tags (comma separated)</label>
-        <input type="text" name="tags" class="form-control" ng-model="gallery.tags" /></textarea>
+        <input type="text" name="tags" class="form-control" ng-model="gallery.tags" placeholder="Tags (comma separated)" />
       </div>
       <div class="form-group">
         <label>Cover</label>
-        <nope-model href="#/media" ng-model="gallery.cover" multiple="false" label="Add gallery cover" preview="icon"></nope-model>
+        <nope-model href="#/media" ng-model="gallery.cover" multiple="false" label="Add gallery cover" preview="icon" template="media"></nope-model>
       </div>
       <div class="form-group">
         <label>Media</label>
-        <nope-model href="#/media?excluded={{(gallery.media | nopeGetIds).join(',')}}" ng-model="gallery.media" label="Add media" preview="icon"></nope-model>
+        <nope-model href="#/media?excluded={{(gallery.media | nopeGetIds).join(',')}}" ng-model="gallery.media" label="Add media" preview="icon" template="media" multiple="true"></nope-model>
       </div>
     </div>
     <div class="panel-footer">